refactor: move test files from src/ to test/
- Move 5 test files + test-utils.ts to test/ directory
- Update tsconfig: rootDir='.', include=['src','test']
- Update package.json: main/bin point to dist/src/, test runs dist/test/
- Published package (files field) only includes dist/src/
