+++ package.json
@@ -3,13 +3,13 @@
   "version": "0.1.0",
   "description": "Local memory management MCP server for Copilot CLI",
   "type": "module",
-  "main": "dist/server.js",
+  "main": "dist/src/server.js",
   "bin": {
-    "memory-mcp": "dist/setup.js",
-    "memory-mcp-cli": "dist/cli.js"
+    "memory-mcp": "dist/src/setup.js",
+    "memory-mcp-cli": "dist/src/cli.js"
   },
   "files": [
-    "dist",
+    "dist/src",
     "README.md",
     "LICENSE"
   ],
@@ -20,7 +20,7 @@
     "prepare": "npm run build",
     "prepublishOnly": "npm run build",
     "preuninstall": "node dist/setup.js uninstall; node dist/setup.js uninstall --claude",
-    "test": "node --test dist/**/*.test.js"
+    "test": "node --test dist/test/**/*.test.js"
   },
   "keywords": [
     "mcp",
